Add all BinaryFormatter APIs to netstandard ref
Expose all the APIs in BinaryFormatter along with the full
closure. This adds a number of remoting interfaces.
